[DP-77] fix: revert app route
