--- src/app/(has-nav)/data/layout.tsx
@@ -1,5 +1,0 @@
-import { ReactNode } from "react";
-
-export default function DataLayout({ children }: { children: ReactNode }) {
-  return <div className="relative h-[100dvh] w-full bg-primary2">{children}</div>;
-}
